[FE] REFACTOR: Rank => Ranking으로 변경
